virtual-idol-demo/requirements.txt
# Core Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-experimental>=0.0.10

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Databases
neo4j>=5.0.0
chromadb>=0.4.0

# Vector and Embeddings
sentence-transformers>=2.2.0

# Frontend
streamlit>=1.31.0
pyvis>=0.3.0
networkx>=3.0

# Tools
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
